Rename classes based on PR feedback
