+++ src/Model/Serializer/ValueObject/Calendar/CalendarNormalizer.php
@@ -4,15 +4,17 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar;
 
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours\AdjustedDayNormalizer;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours\ClosedDayNormalizer;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\BookingAvailability;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\BookingAvailabilityType;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\Calendar;
-use CultuurNet\UDB3\Model\ValueObject\Calendar\CalendarWithOpeningHoursAdjusted;
-use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDay;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\CalendarWithClosedDays;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\CalendarWithDateRange;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\CalendarWithOpeningHours;
+use CultuurNet\UDB3\Model\ValueObject\Calendar\CalendarWithAdjustedDays;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\CalendarWithSubEvents;
+use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDay;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\ClosedDay;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\Status;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\StatusType;
@@ -71,7 +73,7 @@ final class CalendarNormalizer implements NormalizerInterface
             );
         }
 
-        if ($calendar instanceof CalendarWithOpeningHoursAdjusted && !$calendar->getAdjustedDays()->isEmpty()) {
+        if ($calendar instanceof CalendarWithAdjustedDays && !$calendar->getAdjustedDays()->isEmpty()) {
             $adjustedOpeningHoursNormalizer = new AdjustedDayNormalizer();
             $data['openingHoursAdjustedDays'] = array_map(
                 fn (AdjustedDay $aoh) => $adjustedOpeningHoursNormalizer->normalize($aoh),
