+++ src/Model/Serializer/ValueObject/Calendar/OpenHours/AdjustedDaysDenormalizer.php
@@ -2,17 +2,19 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar;
+namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpenHours;
 
 use CultuurNet\UDB3\DateTimeFactory;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHourDenormalizer;
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\TranslatedAdjustedDescriptionDenormalizer;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDay;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDays;
+use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\OpeningHour;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\OpeningHours;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\TranslatedAdjustedDescription;
-use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\OpeningHour;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 
-final class AdjustedDayDenormalizer implements DenormalizerInterface
+final class AdjustedDaysDenormalizer implements DenormalizerInterface
 {
     private OpeningHourDenormalizer $openingHourDenormalizer;
     private TranslatedAdjustedDescriptionDenormalizer $translatedDescriptionDenormalizer;
