+++ src/Model/ValueObject/Calendar/CalendarWithAdjustedDays.php
@@ -6,7 +6,7 @@ namespace CultuurNet\UDB3\Model\ValueObject\Calendar;
 
 use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\AdjustedDays;
 
-interface CalendarWithOpeningHoursAdjusted extends Calendar
+interface CalendarWithAdjustedDays extends Calendar
 {
     public function getAdjustedDays(): AdjustedDays;
